refactor: move enums and functions closer to implementation/usage
