+++ src/beaminteraction/src/4C_beaminteraction_calc_utils.hpp
@@ -10,8 +10,7 @@
 
 #include "4C_config.hpp"
 
-#include "4C_beaminteraction_input.hpp"
-#include "4C_binstrategy_utils.hpp"
+#include "4C_beaminteraction_crosslinking_submodel_evaluator.hpp"
 #include "4C_linalg_fevector.hpp"
 #include "4C_linalg_fixedsizematrix.hpp"
 #include "4C_linalg_mapextractor.hpp"
