+++ src/beaminteraction/src/4C_beaminteraction_data.hpp
@@ -10,6 +10,7 @@
 
 #include "4C_config.hpp"
 
+#include "4C_beaminteraction_crosslinking_submodel_evaluator.hpp"
 #include "4C_beaminteraction_input.hpp"
 #include "4C_linalg_fixedsizematrix.hpp"
