+++ src/beaminteraction/src/4C_beaminteraction_input.hpp
@@ -39,149 +39,6 @@ namespace BeamInteraction
     bounding_volume_hierarchy  ///< coupling pair search based on a bounding volume hierarchy
   };
 
-  /// type of the used submodel for beaminteraction
-  enum SubModelType
-  {
-    submodel_crosslinking,    ///< evaluate the structural model
-    submodel_beamcontact,     ///< evaluate the contact model
-    submodel_potential,       ///< evaluate the model for potential-based interactions
-    submodel_spherebeamlink,  ///< evaluate model for cell filament interactions
-    submodel_vague            ///< undefined model type
-  };
-
-  /// type of employed solving strategy for contact
-  /// (this enum represents the input file parameter STRATEGY)
-  enum Strategy
-  {
-    bstr_none,    ///< no beam contact
-    bstr_penalty  ///< penalty method
-  };
-
-  /// type of linker
-  enum JointType
-  {
-    beam3r_line2_rigid,  ///< rigid joint
-    beam3r_line2_pin,    ///< pin joint
-    truss                ///< truss
-  };
-
-  /// type of filament
-  enum FilamentType
-  {
-    filetype_arbitrary,  ///< no special type
-    filetype_actin,      ///< actin type
-    filetype_collagen,   ///< collagen type
-    filetype_none        ///< filament which does not bind
-  };
-
-  /// type of crosslinker
-  enum CrosslinkerType
-  {
-    linkertype_arbitrary,  ///< binds to all filament
-    linkertype_actin,      ///< only binds to actin filaments
-    linkertype_collagen,   ///< only binds to collagen filaments
-    linkertype_integrin    ///< sphere to beam linker
-  };
-
-  /**
-   * \brief Types of beam interaction conditions
-   */
-  enum class BeamInteractionConditionTypes
-  {
-    //! Default value.
-    none,
-    //! Beam-to-beam contact.
-    beam_to_beam_contact,
-    //! Beam-to-solid volume mesh tying.
-    beam_to_solid_volume_meshtying,
-    //! Beam-to-solid surface mesh tying.
-    beam_to_solid_surface_meshtying,
-    //! beam-to-beam penalty point coupling with direct pair creation.
-    beam_to_beam_point_coupling_direct,
-    //! beam-to-beam penalty point coupling with indirect pair creation.
-    beam_to_beam_point_coupling_indirect,
-    //! Beam-to-solid surface contact.
-    beam_to_solid_surface_contact,
-    //! Beam-to-solid edge contact.
-    beam_to_solid_edge_contact
-  };
-
-  //! Map type std::string to enum
-  inline JointType string_to_joint_type(const std::string& name)
-  {
-    JointType type = beam3r_line2_rigid;
-    if (name == "beam3rline2rigid")
-      type = beam3r_line2_rigid;
-    else if (name == "beam3rline2pin")
-      type = beam3r_line2_pin;
-    else if (name == "truss")
-      type = truss;
-    else
-      FOUR_C_THROW("invalid filament type std::string ");
-
-    return type;
-  };
-
-  //! Map type std::string to enum
-  inline FilamentType string_to_filament_type(const std::string& name)
-  {
-    FilamentType type = filetype_arbitrary;
-    if (name == "arbitrary")
-      type = filetype_arbitrary;
-    else if (name == "actin")
-      type = filetype_actin;
-    else if (name == "collagen")
-      type = filetype_collagen;
-    else if (name == "none")
-      type = filetype_none;
-    else
-      FOUR_C_THROW("invalid filament type std::string ");
-
-    return type;
-  };
-
-  //! Map type std::string to enum
-  inline CrosslinkerType string_to_crosslinker_type(const std::string& name)
-  {
-    CrosslinkerType type = linkertype_arbitrary;
-    if (name == "arbitrary")
-      type = linkertype_arbitrary;
-    else if (name == "actin")
-      type = linkertype_actin;
-    else if (name == "collagen")
-      type = linkertype_collagen;
-    else if (name == "integrin")
-      type = linkertype_integrin;
-    else
-    {
-      FOUR_C_THROW(
-          "invalid crosslinker type {}. Possible values are arbitrary, actin, collagen and "
-          "integrin.",
-          name.c_str());
-    }
-
-    return type;
-  };
-
-  //! Map action type enum to std::string
-  static inline std::string crosslinker_type_to_string(const CrosslinkerType type)
-  {
-    switch (type)
-    {
-      case linkertype_arbitrary:
-        return "arbitrary";
-      case linkertype_actin:
-        return "actin";
-      case linkertype_collagen:
-        return "collagen";
-      case linkertype_integrin:
-        return "integrin";
-      default:
-        return "unknown";
-    }
-    return "";
-  };
-
   /// beam interaction parameters
   std::vector<Core::IO::InputSpec> valid_parameters();
