+++ src/beaminteraction/src/contact/beam_to_solid/4C_beaminteraction_contact_beam_to_solid_input.cpp
@@ -8,7 +8,6 @@
 #include "4C_beaminteraction_contact_beam_to_solid_input.hpp"
 
 #include "4C_beaminteraction_contact_beam_to_solid_edge_params.hpp"
-#include "4C_beaminteraction_input.hpp"
 #include "4C_fem_condition_definition.hpp"
 #include "4C_geometry_pair_input.hpp"
 #include "4C_io_input_spec_builders.hpp"
