+++ src/beaminteraction/src/contact/beam_to_solid/4C_beaminteraction_contact_beam_to_solid_input.hpp
@@ -11,6 +11,7 @@
 
 #include "4C_config.hpp"
 
+#include "4C_beaminteraction_conditions.hpp"
 #include "4C_fem_condition.hpp"
 #include "4C_io_input_spec.hpp"
