+++ src/beaminteraction/src/crosslinking/4C_beaminteraction_crosslinking_link.hpp
@@ -10,7 +10,7 @@
 
 #include "4C_config.hpp"
 
-#include "4C_beaminteraction_input.hpp"
+#include "4C_beaminteraction_crosslinking_submodel_evaluator.hpp"
 #include "4C_comm_parobject.hpp"
 #include "4C_comm_parobjectfactory.hpp"
 #include "4C_linalg_fixedsizematrix.hpp"
