+++ src/beaminteraction/src/spherebeamlinking/4C_beaminteraction_spherebeamlinking_params.hpp
@@ -10,7 +10,7 @@
 
 #include "4C_config.hpp"
 
-#include "4C_beaminteraction_input.hpp"
+#include "4C_beaminteraction_crosslinking_submodel_evaluator.hpp"
 #include "4C_utils_exceptions.hpp"
 
 #include <memory>
