+++ src/mat/4C_mat_crosslinkermat.hpp
@@ -10,7 +10,7 @@
 
 #include "4C_config.hpp"
 
-#include "4C_beaminteraction_input.hpp"
+#include "4C_beaminteraction_crosslinking_submodel_evaluator.hpp"
 #include "4C_comm_parobjectfactory.hpp"
 #include "4C_mat_material_factory.hpp"
 #include "4C_material_base.hpp"
